refactor: decompose durable reconciliation boundaries
