+++ src/daily-cache-core.ts
@@ -4,8 +4,9 @@ import { mkdir, open, readdir, readFile, rename, stat, unlink } from 'fs/promise
 import { join } from 'path'
 import type { DateRange, ProjectSummary } from './types.js'
 import { getMetroraCacheDir } from './product-paths.js'
-import { emptyModelStats, mergeModelStats, sanitizeModels } from './daily-cache-model-detail.js'
+import { sanitizeModels } from './daily-cache-model-detail.js'
 import type { CategoryDayStats, DailyCache, DailyEntry, ModelDayStats, ProjectDayStats, ProviderDaySlice } from './daily-cache-types.js'
+import { mergeDayEntries, setOwn } from './daily-cache-merge.js'
 // v18: Copilot accounting changed; v17 history is adopted untrusted, re-derived where sources survive, and carried otherwise.
 // Bumped to 16: historical per-call cost assignments. Surviving source days
 // re-derive under immutable date-effective settlements; sourceless provider
@@ -78,6 +79,8 @@ const DAILY_CACHE_FILENAME = `daily-cache.v${DAILY_CACHE_VERSION}.json`
 
 
 export type { CategoryDayStats, DailyCache, DailyEntry, ModelDayStats, ProjectDayStats, ProviderDaySlice } from './daily-cache-types.js'
+export { emptyDailyEntry } from './daily-cache-reconciliation.js'
+export { mergeDayEntries } from './daily-cache-merge.js'
 
 function getCacheDir(): string {
   return getMetroraCacheDir()
@@ -100,26 +103,6 @@ export function dailyCachePath(): string {
 
 export function emptyCache(savingsConfigHash = ''): DailyCache {
   return { version: DAILY_CACHE_VERSION, savingsConfigHash, tzKey: currentTzKey(), lastComputedDate: null, days: [], complete: false }
-}
-
-/** Empty primary row used to tombstone a provider slice on a re-derived day. */
-export function emptyDailyEntry(date: string): DailyEntry {
-  return {
-    date,
-    cost: 0,
-    savingsUSD: 0,
-    calls: 0,
-    sessions: 0,
-    inputTokens: 0,
-    outputTokens: 0,
-    cacheReadTokens: 0,
-    cacheWriteTokens: 0,
-    editTurns: 0,
-    oneShotTurns: 0,
-    models: {},
-    categories: {},
-    providers: {},
-  }
 }
 
 export function isMigratableCache(parsed: unknown): parsed is {
@@ -426,145 +409,6 @@ export function applyRetention(days: DailyEntry[], newestDate: string): DailyEnt
   cutoffDate.setUTCDate(cutoffDate.getUTCDate() - DAILY_CACHE_RETENTION_DAYS)
   const cutoff = toDateString(cutoffDate)
   return days.filter(d => d.date >= cutoff)
-}
-
-function hasSliceData(slice: ProviderDaySlice): boolean {
-  return slice.cost > 0 || slice.calls > 0 || (slice.savingsUSD ?? 0) > 0
-}
-
-/// A day from a pre-v5-era cache: day-level totals exist but the providers map
-/// is empty, so nothing can be attributed per provider. Such a day merges
-/// all-or-nothing — filling slices into it would double-count whatever share
-/// of its totals the incoming provider already contributed.
-function isOpaqueDay(day: DailyEntry): boolean {
-  return (day.cost > 0 || day.calls > 0) && Object.keys(day.providers).length === 0
-}
-
-/// Fold one provider's day slice into a day: the providers map, the day-level
-/// totals, and (when the slice carries them — v14+ slices do) the model and
-/// category breakdowns. Skinny slices from pre-v14 caches restore only
-/// cost/calls/savings; the day's other totals simply don't grow. A zero-data
-/// placeholder already present for the provider (a session that started this
-/// day but whose turns all landed on another) only contributes its session
-/// count, deduplicated by max — the same real session may be counted on both
-/// sides.
-function addSliceIntoDay(day: DailyEntry, provider: string, slice: ProviderDaySlice): void {
-  // Reads keyed by names from foreign caches use hasOwn throughout: a plain
-  // lookup of "__proto__" returns the prototype object, and accumulating into
-  // it pollutes every object in the process.
-  const placeholder = Object.hasOwn(day.providers, provider) ? day.providers[provider] : undefined
-  const placeholderSessions = placeholder?.sessions ?? 0
-  const merged = structuredClone(slice)
-  if (placeholderSessions > (merged.sessions ?? 0)) merged.sessions = placeholderSessions
-  setOwn(day.providers, provider, merged)
-  day.cost += slice.cost
-  day.calls += slice.calls
-  day.savingsUSD += slice.savingsUSD ?? 0
-  day.sessions += Math.max(0, (slice.sessions ?? 0) - placeholderSessions)
-  day.inputTokens += slice.inputTokens ?? 0
-  day.outputTokens += slice.outputTokens ?? 0
-  if (slice.reasoningTokens !== undefined) day.reasoningTokens = (day.reasoningTokens ?? 0) + slice.reasoningTokens
-  day.cacheReadTokens += slice.cacheReadTokens ?? 0
-  day.cacheWriteTokens += slice.cacheWriteTokens ?? 0
-  day.editTurns += slice.editTurns ?? 0
-  day.oneShotTurns += slice.oneShotTurns ?? 0
-  for (const [name, m] of Object.entries(slice.models ?? {})) {
-    const acc = Object.hasOwn(day.models, name) ? day.models[name]! : emptyModelStats()
-    mergeModelStats(acc, m)
-    setOwn(day.models, name, acc)
-  }
-  for (const [cat, c] of Object.entries(slice.categories ?? {})) {
-    const acc = Object.hasOwn(day.categories, cat) ? day.categories[cat]! : { turns: 0, cost: 0, savingsUSD: 0, editTurns: 0, oneShotTurns: 0 }
-    acc.turns += c.turns
-    acc.cost += c.cost
-    acc.savingsUSD += c.savingsUSD ?? 0
-    acc.editTurns += c.editTurns
-    acc.oneShotTurns += c.oneShotTurns
-    setOwn(day.categories, cat, acc)
-  }
-  const placeholderProjects = placeholder?.projects ?? {}
-  for (const [name, p] of Object.entries(slice.projects ?? {})) {
-    if (!p || typeof p !== 'object' || Array.isArray(p)) continue
-    const dayProjects = (day.projects ??= {})
-    const acc = Object.hasOwn(dayProjects, name) ? dayProjects[name]! : { cost: 0, calls: 0, savingsUSD: 0, sessions: 0 }
-    acc.cost += num(p.cost)
-    acc.calls += num(p.calls)
-    acc.savingsUSD += num(p.savingsUSD)
-    if (!acc.path && typeof p.path === 'string') acc.path = p.path
-    // Same session dedup as the slice-level sessions above: a placeholder's
-    // project sessions were already counted into the day when the fresh day
-    // was built, so only the excess is added.
-    const placeholderProjectSessions = Object.hasOwn(placeholderProjects, name) ? num(placeholderProjects[name]?.sessions) : 0
-    acc.sessions += Math.max(0, num(p.sessions) - placeholderProjectSessions)
-    setOwn(dayProjects, name, acc)
-  }
-  // Placeholder-only projects (session counted fresh, calls landed elsewhere)
-  // survive on the merged slice rather than being dropped by the clone above.
-  const mergedProjects = merged.projects
-  if (mergedProjects) {
-    for (const [name, p] of Object.entries(placeholderProjects)) {
-      if (!p || typeof p !== 'object') continue
-      if (Object.hasOwn(mergedProjects, name)) {
-        if (num(p.sessions) > num(mergedProjects[name]!.sessions)) mergedProjects[name]!.sessions = num(p.sessions)
-      } else {
-        setOwn(mergedProjects, name, { cost: 0, calls: 0, savingsUSD: 0, sessions: num(p.sessions) })
-      }
-    }
-  } else if (placeholder?.projects) {
-    merged.projects = structuredClone(placeholder.projects)
-  }
-}
-
-/// Assign via defineProperty so filesystem-derived keys like "__proto__" become
-/// ordinary own properties instead of mutating the prototype link.
-function setOwn<T>(target: Record<string, T>, key: string, value: T): void {
-  Object.defineProperty(target, key, { value, enumerable: true, writable: true, configurable: true })
-}
-
-/// Merge two day lists per (date, provider): `primary` wins wherever both have
-/// data; `secondary` only fills dates primary lacks entirely and provider
-/// slices primary lacks on shared dates. Nothing in secondary can overwrite or
-/// double into primary. With markSecondaryCarried, every day that received a
-/// secondary contribution is flagged `carried`.
-///
-/// Days that cannot be attributed per provider merge all-or-nothing:
-///  - an OPAQUE primary day (pre-v5-era: totals but empty providers map) is
-///    never slice-filled — its totals may already contain the incoming
-///    provider's share, so adding would double-count;
-///  - an opaque secondary day on a date primary already has contributes
-///    nothing — its day-level totals cannot be attributed without slices.
-/// A primary slice blocks a secondary one only when it carries DATA; a
-/// zero-data placeholder (sessions only) is merged into, not treated as a
-/// re-derivation of the provider's day.
-export function mergeDayEntries(
-  primary: DailyEntry[],
-  secondary: DailyEntry[],
-  markSecondaryCarried: boolean,
-  blockedSecondaryProviderDays: ReadonlySet<string> = new Set(),
-): DailyEntry[] {
-  const byDate = new Map<string, DailyEntry>()
-  for (const day of primary) byDate.set(day.date, structuredClone(day))
-  for (const day of secondary) {
-    const existing = byDate.get(day.date)
-    if (!existing) {
-      const copy = structuredClone(day)
-      if (markSecondaryCarried) copy.carried = true
-      byDate.set(day.date, copy)
-      continue
-    }
-    if (isOpaqueDay(existing)) continue
-    for (const [provider, slice] of Object.entries(day.providers)) {
-      if (blockedSecondaryProviderDays.has(`${provider}\u0000${day.date}`)) continue
-      // Sessions-only slices (a session whose calls all landed on another
-      // day) still carry a real session count — worth preserving.
-      if (!hasSliceData(slice) && !(slice.sessions ?? 0)) continue
-      const existingSlice = Object.hasOwn(existing.providers, provider) ? existing.providers[provider] : undefined
-      if (existingSlice && hasSliceData(existingSlice)) continue
-      addSliceIntoDay(existing, provider, slice)
-      if (markSecondaryCarried) existing.carried = true
-    }
-  }
-  return [...byDate.values()].sort((a, b) => a.date.localeCompare(b.date))
 }
 
 export function getDaysInRange(cache: DailyCache, start: string, end: string): DailyEntry[] {
